Move RtpCapabilities into separate module
